refactored specs fixes #552
